backend/requirements.txt
# FastAPI and dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.3.0

# Database
supabase==2.0.3
sqlalchemy==2.0.23
redis==6.4.0
psycopg2-binary==2.9.10

# CORS and security
python-jose[cryptography]==3.3.0
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
bcrypt==4.3.0
python-dotenv==1.0.0
itsdangerous==2.2.0

# File handling
aiofiles==23.2.1
openpyxl==3.1.2

# PDF Generation
playwright==1.40.0
weasyprint==62.3
jinja2==3.1.2

# Email (optional, for sending documents)
# python-emails==0.5.15

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.24.1